1) Поправил инкапсуляцию в классе UserServiceImpl. После чего передал объект типа UserDao в аргументы метода. После чего в классе Main создал объект UserDao и передал его в конструктор UserServiceImpl. 2) Исправил dropUserTable, добавил проверку на существование таблицы перед вызовом метода и немного модифицировал сам метод. 3) Определил метод закрытия для объекта Connection в классе Util и вызваю closeConnection() в конце метода main().

+++ src/main/java/jm/task/core/jdbc/dao/UserDaoJDBCImpl.java
@@ -33,14 +33,40 @@ public class UserDaoJDBCImpl implements UserDao {
         }
     }
 
+    // проверка на существование таблицы перед вызовом метода dropUsersTable()
+    public boolean isUsersTableExists() {
+        try {
+            DatabaseMetaData metaData = connection.getMetaData();
+            ResultSet resultSet = metaData.getTables(null, null, "User", null);
+            return resultSet.next();
+        } catch (SQLException e) {
+            logger.log(Level.SEVERE, "Ошибка при проверке существования таблицы", e);
+            return false;
+        }
+    }
+
+//    @Override
+//    public void dropUsersTable() {
+//        String createTableSQL = "DROP TABLE User";
+//
+//        try (Statement statement = connection.createStatement()) {
+//            statement.execute(createTableSQL);
+//        } catch (SQLException e) {
+//            logger.log(Level.SEVERE, "Неудалось удалить таблицу 2 \n", e);
+//        }
+//    }
+
     @Override
     public void dropUsersTable() {
-        String createTableSQL = "DROP TABLE User";
-
-        try (Statement statement = connection.createStatement()) {
-            statement.execute(createTableSQL);
-        } catch (SQLException e) {
-            logger.log(Level.SEVERE, "Неудалось удалить таблицу 2 \n", e);
+        if (isUsersTableExists()) {
+            String createTableSQL = "DROP TABLE User";
+            try (Statement statement = connection.createStatement()) {
+                statement.execute(createTableSQL);
+            } catch (SQLException e) {
+                logger.log(Level.SEVERE, "Неудалось удалить таблицу", e);
+            }
+        } else {
+            logger.info("Таблица пользователей не существует");
         }
     }
